requirements.txt
streamlit==1.28.1
PyMuPDF==1.23.5
pdfplumber==0.10.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
Pillow==10.0.1
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
pdf2image
requests